Remove validate_datarules.cpp
* Checks moved into individual opcode validation
  * removes duplicated checks
* Add check that forward pointer points to struct
